nfc: move iso15 parser to libs
