+++ lib/signal_reader/SConscript
@@ -11,14 +11,7 @@ env.Append(
 
 libenv = env.Clone(FW_LIB_NAME="signal_reader")
 libenv.ApplyLibFlags()
-
-libenv.AppendUnique(
-    CCFLAGS=[
-        # Required for lib to be linkable with .faps
-        # "-mword-relocations",
-        # "-mlong-calls",
-    ],
-)
+libenv.Append(CCFLAGS=["-O3", "-funroll-loops", "-Ofast"])
 
 sources = libenv.GlobRecursive("*.c*")
